Combine style diff into a single version
per https://github.com/mapbox/mapbox-gl-style-spec/pull/324#issuecomment-127820588
