--- diff/latest.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = require('./v7');
