--- diff/v7.js
@@ -1,301 +1,0 @@
-'use strict';
-
-var isEqual = require('lodash.isequal');
-
-var operations = {
-
-    /*
-     * { command: 'setStyle', args: [stylesheet] }
-     */
-    setStyle: 'setStyle',
-
-    /*
-     * { command: 'addLayer', args: [layer, 'beforeLayerId'] }
-     */
-    addLayer: 'addLayer',
-
-    /*
-     * { command: 'removeLayer', args: ['layerId'] }
-     */
-    removeLayer: 'removeLayer',
-
-    /*
-     * { command: 'setPaintProperty', args: ['layerId', 'prop', value] }
-     */
-    setPaintProperty: 'setPaintProperty',
-
-    /*
-     * { command: 'setLayoutProperty', args: ['layerId', 'prop', value] }
-     */
-    setLayoutProperty: 'setLayoutProperty',
-
-    /*
-     * { command: 'setFilter', args: ['layerId', filter] }
-     */
-    setFilter: 'setFilter',
-
-    /*
-     * { command: 'addSource', args: ['sourceId', source] }
-     */
-    addSource: 'addSource',
-
-    /*
-     * { command: 'removeSource', args: ['sourceId'] }
-     */
-    removeSource: 'removeSource',
-
-    /*
-     * { command: 'setLayerZoomRange', args: ['layerId', 0, 22] }
-     */
-    setLayerZoomRange: 'setLayerZoomRange',
-
-    /*
-     * { command: 'setConstant', args: ['constantName', value] }
-     */
-    setConstant: 'setConstant',
-
-    /*
-     * { command: 'setSprite', args: ['spriteUrl'] }
-     */
-    setSprite: 'setSprite',
-
-    /*
-     * { command: 'setGlyphs', args: ['glyphsUrl'] }
-     */
-    setGlyphs: 'setGlyphs',
-
-    /*
-     * { command: 'setTransition', args: [transition] }
-     */
-    setTransition: 'setTransition',
-
-    /*
-     * { command: 'setLayerProperty', args: ['layerId', 'prop', value] }
-     */
-    setLayerProperty: 'setLayerProperty'
-
-};
-
-
-function diffConstants(before, after, commands) {
-    before = before || {};
-    after = after || {};
-
-    var prop;
-
-    for (prop in before) {
-        if (!before.hasOwnProperty(prop)) continue;
-        if (!isEqual(before[prop], after[prop])) {
-            commands.push({ command: operations.setConstant, args: [prop, after[prop]] });
-        }
-    }
-    for (prop in after) {
-        if (!after.hasOwnProperty(prop) || before.hasOwnProperty(prop)) continue;
-        if (!isEqual(before[prop], after[prop])) {
-            commands.push({ command: operations.setConstant, args: [prop, after[prop]] });
-        }
-    }
-}
-
-function diffSources(before, after, commands) {
-    before = before || {};
-    after = after || {};
-
-    var sourceId;
-
-    // look for sources to remove
-    for (sourceId in before) {
-        if (!before.hasOwnProperty(sourceId)) continue;
-        if (!after.hasOwnProperty(sourceId)) {
-            commands.push({ command: operations.removeSource, args: [sourceId] });
-        }
-    }
-
-    // look for sources to add/update
-    for (sourceId in after) {
-        if (!after.hasOwnProperty(sourceId)) continue;
-        if (!before.hasOwnProperty(sourceId)) {
-            commands.push({ command: operations.addSource, args: [sourceId, after[sourceId]] });
-        } else if (!isEqual(before[sourceId], after[sourceId])) {
-            // no update command, must remove then add
-            commands.push({ command: operations.removeSource, args: [sourceId] });
-            commands.push({ command: operations.addSource, args: [sourceId, after[sourceId]] });
-        }
-    }
-}
-
-function diffLayerPropertyChanges(before, after, commands, layerId, klass, command) {
-    before = before || {};
-    after = after || {};
-
-    var prop;
-
-    for (prop in before) {
-        if (!before.hasOwnProperty(prop)) continue;
-        if (!isEqual(before[prop], after[prop])) {
-            commands.push({ command: command, args: [layerId, prop, after[prop], klass] });
-        }
-    }
-    for (prop in after) {
-        if (!after.hasOwnProperty(prop) || before.hasOwnProperty(prop)) continue;
-        if (!isEqual(before[prop], after[prop])) {
-            commands.push({ command: command, args: [layerId, prop, after[prop], klass] });
-        }
-    }
-}
-
-function pluckId(layer) {
-    return layer.id;
-}
-function indexById(group, layer) {
-    group[layer.id] = layer;
-    return group;
-}
-
-function diffLayers(before, after, commands) {
-    before = before || [];
-    after = after || [];
-
-    // order of layers by id
-    var beforeOrder = before.map(pluckId);
-    var afterOrder = after.map(pluckId);
-
-    // index of layer by id
-    var beforeIndex = before.reduce(indexById, {});
-    var afterIndex = after.reduce(indexById, {});
-
-    // track order of layers as if they have been mutated
-    var tracker = beforeOrder.slice();
-
-    // layers that have been added do not need to be diffed
-    var clean = Object.create(null);
-
-    var i, d, layerId, beforeLayer, afterLayer, insertBeforeLayerId, prop;
-
-    // remove layers
-    for (i = 0, d = 0; i < beforeOrder.length; i++) {
-        layerId = beforeOrder[i];
-        if (!afterIndex.hasOwnProperty(layerId)) {
-            commands.push({ command: operations.removeLayer, args: [layerId] });
-            tracker.splice(tracker.indexOf(layerId, d), 1);
-        } else {
-            // limit where in tracker we need to look for a match
-            d++;
-        }
-    }
-
-    // add/reorder layers
-    for (i = 0, d = 0; i < afterOrder.length; i++) {
-        // work backwards as insert is before an existing layer
-        layerId = afterOrder[afterOrder.length - 1 - i];
-
-        if (tracker[tracker.length - 1 - i] === layerId) continue;
-
-        if (beforeIndex.hasOwnProperty(layerId)) {
-            // remove the layer before we insert at the correct position
-            commands.push({ command: operations.removeLayer, args: [layerId] });
-            tracker.splice(tracker.lastIndexOf(layerId, tracker.length - d), 1);
-        } else {
-            // limit where in tracker we need to look for a match
-            d++;
-        }
-
-        // add layer at correct position
-        insertBeforeLayerId = tracker[tracker.length - i];
-        commands.push({ command: operations.addLayer, args: [afterIndex[layerId], insertBeforeLayerId] });
-        tracker.splice(tracker.length - i, 0, layerId);
-        clean[layerId] = true;
-    }
-
-    // update layers
-    for (i = 0; i < afterOrder.length; i++) {
-        layerId = afterOrder[i];
-        beforeLayer = beforeIndex[layerId];
-        afterLayer = afterIndex[layerId];
-
-        // no need to update if previously added (new or moved)
-        if (clean[layerId] || isEqual(beforeLayer, afterLayer)) continue;
-
-        // layout, paint, filter, minzoom, maxzoom
-        diffLayerPropertyChanges(beforeLayer.layout, afterLayer.layout, commands, layerId, null, operations.setLayoutProperty);
-        diffLayerPropertyChanges(beforeLayer.paint, afterLayer.paint, commands, layerId, null, operations.setPaintProperty);
-        if (!isEqual(beforeLayer.filter, afterLayer.filter)) {
-            commands.push({ command: operations.setFilter, args: [layerId, afterLayer.filter] });
-        }
-        if (!isEqual(beforeLayer.minzoom, afterLayer.minzoom) || !isEqual(beforeLayer.maxzoom, afterLayer.maxzoom)) {
-            commands.push({ command: operations.setLayerZoomRange, args: [layerId, afterLayer.minzoom, afterLayer.maxzoom] });
-        }
-
-        // handle all other layer props, including paint.*
-        for (prop in beforeLayer) {
-            if (!beforeLayer.hasOwnProperty(prop)) continue;
-            if (prop === 'layout' || prop === 'paint' || prop === 'filter'
-                || prop === 'metadata' || prop === 'minzoom' || prop === 'maxzoom') continue;
-            if (prop.indexOf('paint.') === 0) {
-                diffLayerPropertyChanges(beforeLayer[prop], afterLayer[prop], commands, layerId, prop.slice(6), operations.setPaintProperty);
-            } else if (!isEqual(beforeLayer[prop], afterLayer[prop])) {
-                commands.push({ command: operations.setLayerProperty, args: [layerId, prop, afterLayer[prop]] });
-            }
-        }
-        for (prop in afterLayer) {
-            if (!afterLayer.hasOwnProperty(prop) || beforeLayer.hasOwnProperty(prop)) continue;
-            if (prop === 'layout' || prop === 'paint' || prop === 'filter'
-                || prop === 'metadata' || prop === 'minzoom' || prop === 'maxzoom') continue;
-            if (prop.indexOf('paint.') === 0) {
-                diffLayerPropertyChanges(beforeLayer[prop], afterLayer[prop], commands, layerId, prop.slice(6), operations.setPaintProperty);
-            } else if (!isEqual(beforeLayer[prop], afterLayer[prop])) {
-                commands.push({ command: operations.setLayerProperty, args: [layerId, prop, afterLayer[prop]] });
-            }
-        }
-    }
-}
-
-/**
- * Diff two stylesheet
- *
- * Creates semanticly aware diffs that can easily be applied at runtime.
- * Operations produced by the diff closely resemble the mapbox-gl-js API. Any
- * error creating the diff will fall back to the 'setStyle' operation.
- *
- * Example diff:
- * [
- *     { command: 'setConstant', args: ['@water', '#0000FF'] },
- *     { command: 'setPaintProperty', args: ['background', 'background-color', 'black'] }
- * ]
- *
- * @param {*} [before] stylesheet to compare from
- * @param {*} after stylesheet to compare to
- * @returns Array list of changes
- */
-function diffStyles(before, after) {
-    if (!before) return [{ command: operations.setStyle, args: [after] }];
-
-    var commands = [];
-
-    try {
-        if (!isEqual(before.version, after.version)) {
-            return [{ command: operations.setStyle, args: [after] }];
-        }
-        if (!isEqual(before.sprite, after.sprite)) {
-            commands.push({ command: operations.setSprite, args: [after.sprite] });
-        }
-        if (!isEqual(before.glyphs, after.glyphs)) {
-            commands.push({ command: operations.setGlyphs, args: [after.glyphs] });
-        }
-        if (!isEqual(before.transition, after.transition)) {
-            commands.push({ command: operations.setTransition, args: [after.transition] });
-        }
-        diffConstants(before.constants, after.constants, commands);
-        diffSources(before.sources, after.sources, commands);
-        diffLayers(before.layers, after.layers, commands);
-    } catch (e) {
-        // fall back to setStyle
-        console.warn('Unable to compute style diff:', e);
-        commands = [{ command: operations.setStyle, args: [after] }];
-    }
-
-    return commands;
-}
-
-module.exports = diffStyles;
-module.exports.operations = operations;
