+++ index.js
@@ -10,4 +10,4 @@ exports.format = require('./lib/format');
 exports.migrate = require('./lib/migrate');
 exports.validate = require('./lib/validate');
 exports.createSprite = require('./lib/create_sprite');
-exports.diff = require('./diff/latest');
+exports.diff = require('./lib/diff');
